flask/requirements.txt
Flask==3.0.3
wikipedia-api==0.6.0
google-generativeai==0.8.3
gunicorn==23.0.0
requests==2.32.3
